Project directories and components refactor
