+++ src/App.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import {createTheme, ThemeProvider} from '@mui/material/styles';
-import Page from "./components/Page";
+import ReportsTabs from "./components/reports/ReportsTabs";
 import {BrowserRouter, Routes, Route} from "react-router-dom";
 import Navbar from "./components/Navbar";
 import Footer from "./components/Footer";
-import SingleReport from "./components/SingleReport";
+import SingleReport from "./components/reports/SingleReport";
 
 const theme = createTheme({
     palette: {
@@ -24,7 +24,7 @@ function App() {
             <BrowserRouter>
                 <Navbar/>
                 <Routes>
-                    <Route path="/" element={<Page/>}></Route>
+                    <Route path="/" element={<ReportsTabs/>}></Route>
                     <Route path="/report/:id" element={<SingleReport/>}></Route>
                 </Routes>
                 <Footer/>
